+++ frontend/package.json
@@ -11,7 +11,6 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@types/styled-components": "^5.1.15",
     "axios": "^0.24.0",
     "core-js": "^3.19.0",
     "dotenv": "^10.0.0",
@@ -23,6 +22,7 @@
     "react-query": "^3.32.3",
     "react-router-dom": "5.2.0",
     "react-spinners": "^0.11.0",
+    "@testing-library/react": "^12.1.2",
     "recoil": "^0.4.1",
     "styled-components": "^5.3.3",
     "styled-tools": "^1.7.2",
@@ -35,7 +35,6 @@
     "@babel/preset-react": "^7.14.5",
     "@babel/preset-typescript": "^7.16.0",
     "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.1.2",
     "@types/dotenv": "^8.2.0",
     "@types/jest": "^27.0.2",
     "@types/lodash": "^4.14.176",
