+++ frontend/src/test/page.test.tsx
@@ -1,35 +1,15 @@
 import React from 'react';
-import { RecoilRoot } from 'recoil';
-import { QueryClient, QueryClientProvider } from 'react-query';
 import ReactDOM from 'react-dom';
-import { BrowserRouter } from 'react-router-dom';
-import { render, cleanup, screen, waitFor } from '@testing-library/react';
+import { cleanup, render, screen, waitFor } from '@testing-library/react';
 import '@testing-library/jest-dom/';
 import 'jest-styled-components';
 import axios from 'axios';
-
-import Login from '@pages/Login';
 import BrowseChannelList from '@organisms/BrowseChannelList';
+import Login from '@pages/Login';
+import DefaultEnvironment from './DefaultEnvironment';
 
 jest.mock('axios');
 const mockedAxios = axios as jest.Mocked<typeof axios>;
-const queryClient = new QueryClient();
-
-const DefaultEnvironment = ({
-  children,
-  initializeState,
-}: {
-  children: JSX.Element;
-  initializeState: any;
-}) => {
-  return (
-    <RecoilRoot initializeState={initializeState ?? null}>
-      <QueryClientProvider client={queryClient}>
-        <BrowserRouter>{children}</BrowserRouter>
-      </QueryClientProvider>
-    </RecoilRoot>
-  );
-};
 
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore
